+++ src/Base/boostpython/PyItems.cpp
@@ -31,7 +31,7 @@ template<typename ItemType>
 struct ItemList_to_pylist_converter {
     static PyObject* convert(const ItemList<ItemType>& items){
         py::list retval;
-        for(int i=0; i < items.size(); ++i){
+        for(size_t i=0; i < items.size(); ++i){
             retval.append(items[i]);
         }
         return py::incref(retval.ptr());
